feat: resolved session token issue for offline submission 🎉
